requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Frontend
streamlit==1.28.0
streamlit-webrtc==0.48.0

# AI
openai==1.3.0

# Maps
folium==0.15.1
geopy==2.4.0

# Utilities
pydantic==2.4.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1

pydantic>=2.0.0
pydantic-settings>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
passlib[bcrypt]>=1.7.4